# api/v3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: miniyu97@gmail.com
    name: meteormin
    url: https://github.com/meteormin/friday.go
  description: Friday.go API
  title: Friday.go API
  version: "{{.Version}}"
servers:
- url: "http://{{.Host}}/"
- url: "https://{{.Host}}/"
...
paths:
  /api/auth/has-admin:
    get:
      description: 회원 관리자 여부 조회 API
      operationId: has-admin
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
          description: 회원 관리자 여부 조회 성공
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      summary: 회원 관리자 여부 조회
      tags:
      - auth
  /api/auth/me:
    get:
      description: 회원 정보 조회 API
      operationId: me
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.UserResource'
          description: 회원 정보 조회 성공
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 로그인 정보 없음
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 회원 정보 조회
      tags:
      - auth
  /api/auth/sign-in:
    post:
      description: 회원 로그인 API
      operationId: sign-in
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapter_rest_handler.SignInRequest'
        description: 회원 로그인 정보
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.TokenResource'
          description: 회원 로그인 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 로그인 실험
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      summary: 회원 로그인
      tags:
      - auth
      x-codegen-request-body-name: req
  /api/auth/sign-up:
    post:
      description: 회원 가입 API
      operationId: sign-up
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapter_rest_handler.SignupRequest'
        description: 회원 가입 정보
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.UserResource'
          description: 회원 가입 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청" app.ErrInvalidUserPassword
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 이메일 중복" app.ErrDuplicateUserUsername
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      summary: 회원 가입
      tags:
      - auth
      x-codegen-request-body-name: req
...
  /api/users:
    get:
      description: 회원 리스트 조회 API
      operationId: users
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_adapter_rest_handler.UserResource'
                type: array
          description: 회원 리스트 조회 성공
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      summary: 회원 리스트 조회
      tags:
      - users
components:
  schemas:
    github_com_meteormin_friday_go_internal_app.Error:
      example:
        code: 0
        message: message
        title: title
      properties:
        code:
          type: integer
        message:
          type: string
        title:
          type: string
      type: object
...
    internal_adapter_rest_handler.SignInRequest:
      description: 로그인 요청
      properties:
        password:
          type: string
        username:
          type: string
      type: object
    internal_adapter_rest_handler.SignupRequest:
      description: 가입 요청
      properties:
        id:
          type: integer
        name:
          type: string
        password:
          type: string
        username:
          type: string
      type: object
...
    internal_adapter_rest_handler.TokenResource:
      description: 토큰 정보 리소스
      example:
        exp: 0
        token: token
      properties:
        exp:
          type: integer
        token:
          type: string
      type: object
...
    internal_adapter_rest_handler.UserResource:
      description: 회원 정보 리소스
      example:
        createdAt: createdAt
        name: name
        id: 0
        updatedAt: updatedAt
        username: username
      properties:
        createdAt:
          type: string
        id:
          type: integer
        name:
          type: string
        updatedAt:
          type: string
        username:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...